✨ Add support for the `Wannier90OptimizeWorkChain`
Add support for the (projectability-disentangled Wannier function) PDWF method by
extending the `w90_bands` input port to the inputs of the
`Wannier90OptimizeWorkChain`.

The current approach for figuring out inside the work chain which wannier90 work chain
variant to run is somewhat hacky: it will simply check if the `projwfc` namespace is in
the inputs. In this case, a `Wannier90BandsWorkChain` is run, else the
`Wannier90OptimizeWorkChain`.

Instead of calling the `chk2ukk` Julia script in the `append_text` of _all_ the
`Wannier90Calculation`s of the wrapper wannier90 work chain, we do so in the
`prepend_text` of the `EpwCalculation`. This also makes more sense for the
`Wannier90BandsWorkChain`, as now the postprocess wannier90 was also calling the
`chk2ukk` script, which would fail and lead to error messages in the logs.