✨ Add the `chk2ukk` Julia script
To make it easier (or rather _possible_) for new users to run the `EpwWorkChain`, add
the `chk2ukk.jl` script to the repository, along with a `Project.toml` and
`Manifest.toml` file for setting up a project to run the script in. Some basic
instructions are also added as a `README.md`.